% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcDiagnostics.R
\name{calcDiagnostics}
\alias{calcDiagnostics}
\title{Calculate diagnostics of a Stan occupancy fit}
\usage{
calcDiagnostics(
  occu_data,
  fit,
  test_seasons = NULL,
  verbose = TRUE,
  collapse_AUC = FALSE
)
}
\arguments{
\item{occu_data}{List. A list with two dataframes: one for site data named "site" and one
for visit data named "visit". The site data frame must contain at least one column named
"site_id" indexing the sites and one named "season_id" indexing the seasons. The
visit data frame must at least have one column named "det" with a 1 for detections and
0 for non-detections and a "visit_id" column indexing the visits (overall, not within seasons).}

\item{fit}{A fitted Stan model.}

\item{test_seasons}{A numeric vector containing the seasons (\code{season_id}) used for testing.}

\item{verbose}{Logical. If TRUE (default), progress of diagnostics is displayed.}

\item{collapse_AUC}{Logical, indicating whether season-level AUC Pr(≥1 detection in season t) or
visit-level AUC Pr(detection in visit j) is computed}
}
\value{
A list of diagnostics
}
\description{
Calculate diagnostics of a Stan occupancy fit
}
\examples{
\dontrun{
    calcDiagnostics(occu_data = my_occu_data,
                    fit = my_occu_fit,
                    test_seasons = 10:15
    )
}
}
