% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configWorkflow.R
\name{configWorkflow}
\alias{configWorkflow}
\title{Configure a Workflow for Occupancy Modelling with Stan}
\usage{
configWorkflow(
  data_dir,
  out_dir,
  seasons,
  train_seasons,
  data_trim = c("none", "end", "start"),
  site_ids,
  visit_ids,
  site_covts,
  visit_covts,
  stan_model_dir,
  stan_model_file = NULL,
  param_to_plot,
  param_extra,
  dynamic = TRUE,
  spatial = FALSE,
  occu_temp_re = c("none", "AR1"),
  det_temp_re = c("none", "AR1"),
  prefile = NULL,
  postfile = NULL,
  occu_models,
  phi_models = NULL,
  gamma_models = NULL,
  det_models,
  verbose = FALSE
)
}
\arguments{
\item{data_dir}{Character. The directory where data should be looked for.}

\item{out_dir}{Character. The output directory where results will be saved.}

\item{seasons}{Numeric vector. The set of seasons for the analysis.}

\item{train_seasons}{Numeric vector. The subset of seasons used for model training.}

\item{data_trim}{Character. Trimming method for data: \code{"none"}, \code{"end"}, or \code{"start"}.}

\item{site_ids}{Character vector. Variables that serve as unique identifiers for sites.}

\item{visit_ids}{Character vector. Variables that serve as unique identifiers for visits.}

\item{site_covts}{Character vector. Site-level covariates.}

\item{visit_covts}{Character vector. Visit-level covariates.}

\item{stan_model_dir}{Character. Path to the Stan model directory.}

\item{stan_model_file}{Character. Name of the Stan model file (optional).}

\item{param_to_plot}{Character vector. Parameters to plot in diagnostics.}

\item{param_extra}{Character vector. Additional parameters to monitor.}

\item{dynamic}{Logical. If \code{TRUE} (default), uses a dynamic occupancy model.}

\item{spatial}{Logical. If \code{TRUE}, uses a model with spatial dependency based on a distance matrix.
Defaults to \code{FALSE}.}

\item{occu_temp_re}{Character. Temporal random effect structure for occupancy component: \code{"none"} or \code{"AR1"}.}

\item{det_temp_re}{Character. Temporal random effect structure for detection component: \code{"none"} or \code{"AR1"}}

\item{prefile}{Character. String used as a prefix for naming files (e.g. model fits, diagnostics) (optional).}

\item{postfile}{Character. String used as a postfix for naming files (e.g. model fits, diagnostics) (optional).}

\item{occu_models}{List of formulas. Occupancy model formula/s.}

\item{phi_models}{List of formulas. Survival (persistence) model formula/s (for dynamic models).}

\item{gamma_models}{List of formulas. Colonization model formula/s (for dynamic models).}

\item{det_models}{List of formulas. Detection model formula/s.}

\item{verbose}{Logical. If TRUE all configuration parameters are printed. Defaults to FALSE.}
}
\value{
A list containing the workflow configuration with the following components:
\itemize{
\item \code{data_dir}: Data directory.
\item \code{out_dir}: Output directory.
\item \code{seasons}: seasons included.
\item \code{train_seasons}: Training seasons.
\item \code{data_trim}: Data trimming method.
\item \code{site_ids}: Site identifiers.
\item \code{visit_ids}: Visit identifiers.
\item \code{site_covts}: Site covariates.
\item \code{visit_covts}: Visit covariates.
\item \code{stan_model}: Stan model file path.
\item \code{param_to_plot}: Parameters for plotting.
\item \code{param_extra}: Extra parameters.
\item \code{dynamic}: Dynamic model flag.
\item \code{spatial}: Spatial model flag.
\item \code{prefile}: File prefix.
\item \code{postfile}: File postfix.
\item \code{occu_models}: Occupancy models.
\item \code{phi_models}: Persistence models.
\item \code{gamma_models}: Colonization models.
\item \code{det_models}: Detection models.
\item \code{raw_data_file}: Raw data file path.
\item \code{clean_data_file}: Clean and standardized data file path.
\item \code{occu_data_file}: Occupancy data list file path.
\item \code{test_data_file}: Occupancy test data list path.
\item \code{test_fit_file}: Occupancy test-data fit file path.
\item \code{test_diags_file}: Occupancy test-data fit diagnostics file path.
\item \code{test_sims_file}: File path to occupancy estimates/predictions from test-data fit.
}
}
\description{
This function sets up the configuration for an occupancy modelling workflow,
specifying directories, species selection, seasons, covariates, and model details.
}
\details{
The lists \code{occu_models}, \code{phi_models}, \code{gamma_models} (if they exist) must have the same
length. In the case of dynamic models \code{occu_models} will be used for initial
occupancy, \code{phi_models} will be used for persistance and \code{gamma_models} will be
used for colonization. Models will be paired elementwise, ie., \code{occu_models[[i]]},
\code{phi_models[[i]]}, and \code{gamma_models[[i]]}, for any \code{i}, will be used for the same dynamic
occupancy model. If the model is not dynamic, then \code{occu_models[[i]]} will be
used for occupancy, and the rest will be ignored. Models will be named according
to the names of \code{occu_models} or by a numeric sequence if \code{occu_models} is unnamed.

Temporal random effects defined by \code{occu_temp_re} affect colonization and survival
components of dynamic models and occupancy component of non-dynamic models. Both
\code{occu_temp_re} and \code{det_temp_re} are ignored if a \code{stan_model_file} is defined.
}
\examples{
\dontrun{
# Example configuration
config <- configWorkflow(
  outdir = "results/species1",
  seasons = 2010:2020,
  trainseasons = 2010:2018,
  datatrim = "none",
  site_ids = sites$site_id,
  visit_ids = visits$visit_id,
  site_covts = site_covariates,
  visit_covts = visit_covariates,
  stanmodel = "occupancy.stan",
  paramtoplot = c("psi", "p"),
  paramextra = c("gamma", "phi"),
  dynamic = TRUE,
  spatial = FALSE,
  prefile = "preprocess.R",
  postfile = "postprocess.R",
  occu_models = c("~1", "~covariate1"),
  phi_models = c("~1", "~covariate2"),
  gamma_models = c("~1", "~covariate3"),
  det_models = c("~1", "~covariate4")
)
}

}
