% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDirs.R
\name{createDirs}
\alias{createDirs}
\title{Create directory structure}
\usage{
createDirs(project_dir = getwd())
}
\arguments{
\item{project_dir}{Character. The path where the structure should be created. Defaults to the current working directory.}
}
\description{
This function creates the directory structure for an occupancy modelling workflow
}
\details{
This function is meant to be run at an empty directory to start working on a workflow.
It will check if any of the directories exist and won't create these to avoid losing
existing data.
}
\examples{
\dontrun{
# Example configuration
createDirs()
createDirs("my/preferred/path")
}

}
