% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createNNdistanceMatrix.R
\name{createNNdistanceMatrix}
\alias{createNNdistanceMatrix}
\title{Create Nearest Neighbours distance matrix}
\usage{
createNNdistanceMatrix(dist_matrix, nNN = 24)
}
\arguments{
\item{dist_matrix}{Matrix. A distance matrix where each entry \link{i, j} corresponds to the distance between sites i and j.}

\item{nNN}{Numeric. The number of nearest neighbours to retain. Defaults to 24.}
}
\value{
A distance matrix with the same rows as \code{dist_matrix} and \code{nNN} columns
}
\description{
Create Nearest Neighbours distance matrix
}
\examples{
\dontrun{
createNNdistanceMatrix(dist_matrix, 16)
}
}
