% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillMissingVisits.R
\name{fillMissingVisits}
\alias{fillMissingVisits}
\title{Fill in missing visits in detection data}
\usage{
fillMissingVisits(
  visit_data,
  site_var,
  season_var,
  seasons = NULL,
  period = c("all", "start", "end")
)
}
\arguments{
\item{visit_data}{A detection/non-detection dataframe.}

\item{site_var}{Character string. Name of the variable used for identifying sites}

\item{season_var}{Character string. Name of the variable used for identifying seasons}

\item{seasons}{An integer vector identifying the seasons the data should cover.}

\item{period}{Character string. One of "all" (default), "start" and "end". Defining
which end of the data should be filled with missing values.}
}
\value{
A dataframe with all possible combinations of \code{site_var} and \code{season_var}
values. \code{season_var} values will be taken from \code{season} if specified or from
\code{range(visit_data)} otherwise. Those cases that were not observed will have NA in detection and any
other variables, such as covariates.
}
\description{
Fill in missing visits in detection data
}
\examples{
\dontrun{

visit_data <- fillMissingVisits(visit_data,
                                site_var = "UTM",
                                season_var = "year",
                                seasons = 1990:2020,
                                period = "all")
}
}
