% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeStanDataBundle.R
\name{formToDesignMatrix}
\alias{formToDesignMatrix}
\title{Construct a design matrix from a formula object}
\usage{
formToDesignMatrix(form, df)
}
\arguments{
\item{form}{A formula object}

\item{df}{A data frame with the variables referred to by \code{form}}
}
\value{
A list with design matrices containing the variables
}
\description{
Construct a design matrix from a formula object
}
\details{
The notation "(1|x)" creates a design matrix for random intercepts.
At the moment there are no random coefficients other than intercepts.
}
\examples{
df <- data.frame(x = runif(20, 0, 10),
                 y = runif(20, -10, 0),
                 z = rnorm(20, 0, 1))

form1 <- ~ x + y

form2 <- ~ x * y # (this does not produce an interaction!)

form3 <- ~ X + y + (1|z)

formToDesignMatrix(form1, df)

formToDesignMatrix(form2, df)

formToDesignMatrix(form3, df) # This separates fixed from random effects
}
