% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_sum_exp.R
\name{log_sum_exp}
\alias{log_sum_exp}
\title{Calculate log sum of exponentials}
\usage{
log_sum_exp(mu)
}
\arguments{
\item{mu}{is a vector of values in the log scale. They typically correspond to
log likelihood values for a Bayesian model.}
}
\value{
The log of the sum of exponentials
}
\description{
Calculate log sum of exponentials
}
\details{
The log sum of exponentials is calculated as:
log-sum-exp_m=1^M mu_m = max(mu) + log sum_m^M exp(mu_m - max(mu)),
where mu_m are typically log likelihood values for a Bayesian model obtained
for iteration m of an MCMC algorithm with a total of M iterations.
This prevents underflow. See https://mc-stan.org/docs/stan-users-guide/
for more information
}
\examples{
log_sum_exp(log(5))

}
