% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeStanDataBundle.R
\name{makeStanDataBundle}
\alias{makeStanDataBundle}
\title{Create a Stan data bundle}
\usage{
makeStanDataBundle(
  occu_data,
  occ_formula,
  phi_formula = NULL,
  gamma_formula = NULL,
  det_formula,
  dynamic = FALSE
)
}
\arguments{
\item{occu_data}{List. A list with two dataframes: one for site data named "site" and one
for visit data named "visit". The site data frame must contain at least one column named
"site_id" indexing the sites and one named "season_id" indexing the seasons. The
visit data frame must at least have one column named "det" with a 1 for detections and
0 for non-detections and a "visit_id" column indexing the visits (overall, not within seasons).}

\item{occ_formula}{Formula. A formula object specifying the initial occurrence model}

\item{phi_formula}{Formula. A formula object specifying the persistence dynamics model}

\item{gamma_formula}{Formula. A formula object specifying the colonization dynamics model}

\item{det_formula}{Formula. A formula object specifying the detection model}

\item{dynamic}{Logical. Indicates whether the specified model is a dynamic one (TRUE)
or not (FALSE, default)}
}
\value{
A data bundle to pass on to a Stan model
}
\description{
Create a Stan data bundle
}
\examples{

occu_data <- list(site = data.frame(site_id = rep(1:3, 3),
                                    season_id = rep(1:3, each = 3)),
                  visit = data.frame(site_id = rep(1:3, 7),
                                     season_id = rep(1:3, each = 7),
                                     det = rbinom(21, 1, 0.3),
                                     effort = rpois(21, 5)))

occu_form <- ~ 1
det_form <- ~ effort

makeStanDataBundle(occu_data,
                   occ_formula = occu_form,
                   det_formula = det_form,
                   dynamic = FALSE)
}
