% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseSims.R
\name{summariseSims}
\alias{summariseSims}
\title{Summarise Stan estimates/predictions}
\usage{
summariseSims(
  occu_data,
  fit,
  group_var = c("season_id", "site_id"),
  pred_var = c("psi", "p", "p_cond")
)
}
\arguments{
\item{occu_data}{List. A list with two dataframes: one for site data named "site" and one
for visit data named "visit". The site data frame must contain at least one column named
"site_id" indexing the sites and one named "season_id" indexing the seasons. The
visit data frame must at least have one column named "det" with a 1 for detections and
0 for non-detections and a "visit_id" column indexing the visits (overall, not within seasons).}

\item{pred_var}{Character. The parameter we want to summarise. One of:
"psi", "p", "p_cond", "psi_eq"}

\item{group_vars}{Character. Variable used for grouping predictions. Either \code{season_id} or \code{site_id}.}
}
\value{
A matrix of summarised estimates/predictions by the variables specified in
\code{group_vars}.
}
\description{
Summarise Stan estimates/predictions
}
\examples{
\dontrun{
    summariseSims(occu_data = my_occu_data,
                  fit = my_occu_fit,
                  group_var = "season_id",
                  pred_var = "psi"
    )
}
}
